+++ test/VM/LogicOpsTest.sol.ts
@@ -3,7 +3,7 @@ import { ethers } from "hardhat";
 import { op, AllStandardOps } from "../../utils";
 import { concat, hexlify } from "ethers/lib/utils";
 import {
-  callSize,
+  zipmapSize,
   max_uint256,
   paddedUInt256,
   paddedUInt32,
@@ -80,7 +80,7 @@ describe("LogicOps Test", async function () {
       concat([
           vReport,
           vReportMax,
-        op(Opcode.ZIPMAP, callSize(1, 3, 1)),
+        op(Opcode.ZIPMAP, zipmapSize(1, 3, 1)),
       ]);
 
     await logic.initialize({ sources: [SOURCE(), ZIPMAP_FN()], constants });
