+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "include": ["./test", "utils"],
+  "include": ["test", "utils"],
   "exclude": ["node_modules/*"],
   "compilerOptions": {
     "target": "ES2020",
