+++ utils/constants.ts
@@ -1,8 +1,5 @@
 import { ethers } from "hardhat";
 
-export const CREATOR_FUNDS_RELEASE_TIMEOUT_TESTING = 100;
-export const MAX_RAISE_DURATION_TESTING = 100;
-
 export const smartPoolManagerAddress = process.env.BALANCER_SMART_POOL_MANAGER;
 if (smartPoolManagerAddress) {
   console.log(`using existing SmartPoolManager: ${smartPoolManagerAddress}`);
@@ -37,13 +34,8 @@ export const sixteenZeros = "0000000000000000";
 export const ONE = ethers.BigNumber.from("1" + eighteenZeros);
 export const RESERVE_ONE = ethers.BigNumber.from("1" + sixZeros);
 
-export const RESERVE_MIN_BALANCE = ethers.BigNumber.from("1" + sixZeros);
-
 export const max_uint256 = ethers.BigNumber.from(
   "0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF"
 );
 export const max_uint32 = ethers.BigNumber.from("0xffffffff");
 export const max_uint16 = ethers.BigNumber.from("0xffff");
-
-export const ALWAYS = 0;
-export const NEVER = max_uint256;
