+++ utils/deploy/basic.ts
@@ -1,0 +1,13 @@
+import { ethers } from "hardhat";
+
+export const basicDeploy = async (name, libs) => {
+  const factory = await ethers.getContractFactory(name, {
+    libraries: libs,
+  });
+
+  const contract = await factory.deploy();
+
+  await contract.deployed();
+
+  return contract;
+};
