+++ utils/hardhat.ts
@@ -1,0 +1,13 @@
+import { ethers } from "hardhat";
+
+export const createEmptyBlock = async (count?: number): Promise<void> => {
+  const signers = await ethers.getSigners();
+  const tx = { to: signers[1].address };
+  if (count > 0) {
+    for (let i = 0; i < count; i++) {
+      await signers[0].sendTransaction(tx);
+    }
+  } else {
+    await signers[0].sendTransaction(tx);
+  }
+};
