+++ utils/index.ts
@@ -1,6 +1,9 @@
+export * from "./bytes";
 export * from "./constants";
-export * from "./helpers";
-export * from "./interfaces";
-export * from "./pure-helpers";
-export * from "./typechain-helpers";
-export * from "./types";
+export * from "./deploy";
+export * from "./events";
+export * from "./hardhat";
+export * from "./math";
+export * from "./rainvm";
+export * from "./test";
+export * from "./tier";
