--- utils/interfaces.ts
@@ -1,6 +1,0 @@
-import { Contract } from "ethers";
-import { RedeemableERC20Factory } from "../typechain/RedeemableERC20Factory";
-
-export interface Factories {
-  redeemableERC20Factory: RedeemableERC20Factory & Contract;
-}
